Added chapters to the video table and fetching metadata
We now have the ability to fetch all of the video's metadata using
YoutubeDL in order to populate the information that was added previously
to the videos table.

While we were fetching metadata the chapters bit caught our attention
and would make a great addition to the system, so chapters were added to
the database as well.